gstreamer: Reset PE/PV to fix package install/upgrade issues
Due to gstreamer-plugins-base being a SIGGEN_EXCLUDE_SAFE_RECIPE_DEPS and also
a git based recipe, the PV can get slightly out of sync if a new version of
the recipe is created while also adding a new machine to the build.

Existing machines will get PV = "version+git1+hash", but a new machine will
end up with "version+git0+hash".  Any recipes that depend on this recipe will
have the embedded dependency of only the first build machine, causing
subsequent (usually new) machines to fail to install into the rootfs with
an error about a missing version of a package.

For instance:

  - nothing provides libgstvideo-1.0-0 >= 1.20.5+git1+1f43f76e19 needed by libopencv-videoio406-4.6.0-r0.3.cortexa72_cortexa53
  - nothing provides libgstaudio-1.0-0 >= 1.20.5+git1+1f43f76e19 needed by libopencv-videoio406-4.6.0-r0.3.cortexa72_cortexa53
  - nothing provides libgstpbutils-1.0-0 >= 1.20.5+git1+1f43f76e19 needed by libopencv-videoio406-4.6.0-r0.3.cortexa72_cortexa53
  - nothing provides libgstapp-1.0-0 >= 1.20.5+git1+1f43f76e19 needed by libopencv-videoio406-4.6.0-r0.3.cortexa72_cortexa53
  - nothing provides libgstriff-1.0-0 >= 1.20.5+git1+1f43f76e19 needed by libopencv-videoio406-4.6.0-r0.3.cortexa72_cortexa53

By setting the PE to '1', we reset the recipe version comparison.  Then the
remaining change, PACKAGE_ARCH set in the task do_packagedata to 'all' will
ensure that the PV will be consistent for all machines that are built.

Note, there is a side effect where the PR will now increment across machines.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>