archconfig: Refactor recipe, service and script
This change modifies how archconfig works by moving the configuration to
a pkg_postinst_ontarget() function in the recipe file. This code is
executed on first boot to configure the various package manager related
files in /etc for the specific carrier card and SOM variant in use.

The Systemd service now only checks that the files have been
configured on subsequent boots so that the user is notified of errors.

The change also cleans up the formatting and variables in all files.

There are two potential issues with this implementation
- the scripts rely on freeipmi to detect the SOM and card variant - if
  the ipmi data read from the eeprom is invalid as it might be in some
  of the older board or currently in qemu this will not work
- The pkg_postinst_ontarget() function creates a script
  /etc/rpm-postinsts/100-archconfig which is run on first boot but
  errors are not propogated to the user and the script is deleted once
  it has been run (even if not sucessful) making it difficult for the
  user to see what _should_ have been done if they are tracing back
  issues.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>